--- operators/cmd/crownlabs-operator/test-tenant.yaml
@@ -1,11 +1,0 @@
-apiVersion: crownlabs.polito.it/v1alpha2
-kind: Tenant 
-metadata:
-  name: test-tenant-2
-  labels:
-    crownlabs.polito.it/operator-selector: new  # Match con il target-label dell'operatore
-spec:
-  firstName: Demo
-  lastName: User
-  email: danielebruscia.oppo@gmail.com
-  lastLogin: "2025-05-20T10:00:00Z"
