+++ operators/cmd/operator/setup_keycloak.md
