+++ operators/pkg/controller/tenant/metrics.go
