+++ operators/pkg/controller/tenant/namespaces/namespace_manager.go
