+++ operators/pkg/controller/tenant/namespaces/namespace_manager_test.go
