+++ operators/pkg/controller/utils/auth.go
